test(hooks): fix ModuleWipeHooksTest stub namespace
