+++ tests/Modules/Core/GetModuleInfoTest.php
@@ -11,19 +11,20 @@ namespace {
     }
 }
 
-namespace Lotgd\Tests\Modules {
+namespace Lotgd\Tests\Modules\Core {
     use PHPUnit\Framework\TestCase;
 
     /**
      * @runTestsInSeparateProcesses
      * @preserveGlobalState disabled
+     * @group core
      */
     final class GetModuleInfoTest extends TestCase
     {
         protected function setUp(): void
         {
             if (!function_exists(__NAMESPACE__ . '\\get_module_info')) {
-                $code = file_get_contents(dirname(__DIR__, 2) . '/lib/modules.php');
+                $code = file_get_contents(dirname(__DIR__, 3) . '/lib/modules.php');
                 $code = preg_replace('/^<\\?php\\s*declare\\(strict_types=1\\);\\s*/', '', $code);
                 eval('namespace ' . __NAMESPACE__ . '; ' . $code);
             }
