+++ tests/Modules/Hooks/ModuleHookRegistrationTest.php
@@ -66,7 +66,7 @@ namespace Lotgd\Tests\Modules\Hooks {
     }
 }
 
-namespace Lotgd\Tests\Modules {
+namespace Lotgd\Tests\Modules\Hooks {
 
 use Lotgd\Tests\Stubs\HookHandler;
 use PHPUnit\Framework\TestCase;
@@ -74,6 +74,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * @runTestsInSeparateProcesses
  * @preserveGlobalState disabled
+ * @group hooks
  */
 final class ModuleHookRegistrationTest extends TestCase
 {
