+++ tests/Modules/Hooks/ModuleWipeHooksTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Lotgd\Tests\Modules\ModuleWipeHooks\Stubs {
+namespace Lotgd\Tests\Modules\Hooks\ModuleWipeHooks\Stubs {
     class HookHandler
     {
         public static bool $wiped = false;
@@ -26,14 +26,15 @@ namespace Lotgd\Tests\Modules\Hooks {
     }
 }
 
-namespace Lotgd\Tests\Modules {
+namespace Lotgd\Tests\Modules\Hooks {
 
-use Lotgd\Tests\Modules\ModuleWipeHooks\Stubs\HookHandler;
+use Lotgd\Tests\Modules\Hooks\ModuleWipeHooks\Stubs\HookHandler;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @runTestsInSeparateProcesses
  * @preserveGlobalState disabled
+ * @group hooks
  */
 final class ModuleWipeHooksTest extends TestCase
 {
