+++ tests/Modules/Injection/MassModulePrepareTest.php
@@ -2,19 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Lotgd\Tests\Modules;
+namespace Lotgd\Tests\Modules\Injection;
 
 use PHPUnit\Framework\TestCase;
 
 /**
  * @runTestsInSeparateProcesses
  * @preserveGlobalState disabled
+ * @group injection
  */
 final class MassModulePrepareTest extends TestCase
 {
     public function testMassModulePrepareDelegatesHooks(): void
     {
-        require __DIR__ . '/../Stubs/MassModulePrepareFunctions.php';
+        require __DIR__ . '/../../Stubs/MassModulePrepareFunctions.php';
         \Lotgd\Modules\HookHandler::$received = [];
         \Lotgd\Modules\HookHandler::$calls    = 0;
 
@@ -28,7 +29,7 @@ final class MassModulePrepareTest extends TestCase
 
     public function testMassModulePrepareWithEmptyHooksReturnsTrue(): void
     {
-        require __DIR__ . '/../Stubs/MassModulePrepareFunctions.php';
+        require __DIR__ . '/../../Stubs/MassModulePrepareFunctions.php';
         \Lotgd\Modules\HookHandler::$received = [];
         \Lotgd\Modules\HookHandler::$calls    = 0;
