+++ tests/Modules/Prefs/ModuleObjPrefsTest.php
@@ -29,13 +29,16 @@ namespace {
     }
 }
 
-namespace Lotgd\Tests\Modules {
+namespace Lotgd\Tests\Modules\Prefs {
 
 use Lotgd\Tests\Stubs\Database;
 use Lotgd\Tests\Stubs\DoctrineConnection;
 use Lotgd\Tests\Stubs\DoctrineResult;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * @group prefs
+ */
 final class ModuleObjPrefsTest extends TestCase
 {
     protected function setUp(): void
