+++ tests/Modules/Settings/ModuleSettingsTest.php
@@ -35,11 +35,14 @@ namespace {
     }
 }
 
-namespace Lotgd\Tests\Modules {
+namespace Lotgd\Tests\Modules\Settings {
 
 use Lotgd\Tests\Stubs\Database;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * @group settings
+ */
 final class ModuleSettingsTest extends TestCase
 {
     protected function setUp(): void
